backend/requirements.txt
flask
flask-cors
requests
python-dotenv
svgwrite
gunicorn
